+++ src/cli/update.js
@@ -1,0 +1,8 @@
+export default ({ commandProcessor, root }) => {
+	commandProcessor.createCommand(root, 'update', 'Update the system firmware of a device via USB', {
+		handler: (args) => {
+			const UpdateCommand = require('../cmd/update');
+			return new UpdateCommand(args).updateDevice();
+		}
+	});
+};
